Refactor service imports: consolidate `services` folder into singular `service` directory and revert middleware authentication logic
