+++ src/lib/domain/clan-management/service/index.ts
@@ -1,5 +1,5 @@
-import {ClanarinaService} from "@/lib/domain/clan-management/services/ClanarinaService";
-import {ClanService} from "@/lib/domain/clan-management/services/ClanService";
+import {ClanarinaService} from "@/lib/domain/clan-management/service/ClanarinaService";
+import {ClanService} from "@/lib/domain/clan-management/service/ClanService";
 import {clanarinaRepository, clanRepository} from "@/lib/domain/clan-management/repository";
 
 export const clanarinaService = new ClanarinaService(clanarinaRepository)
