+++ src/middleware.ts
@@ -1,5 +1,4 @@
 import {NextRequest, NextResponse} from 'next/server';
-import {userService} from "@/lib/domain/user-management/service";
 
 export function middleware(request: NextRequest) {
   const { pathname } = request.nextUrl;
@@ -33,11 +32,9 @@ export function middleware(request: NextRequest) {
 
   if (isProtectedPath) {
     // Check for authentication cookie
-      const adminCookie = request.cookies.get('gh_admin');
-      const loggedInUser = request.cookies.get('logged_in_user');
-      const username = loggedInUser?.value;
-      const user = username ? userService.getUserByUsername(username) : null;
-      if ((user && !isUserAuthorized(user)) || !adminCookie || adminCookie.value !== '1') {
+      const authCookie = request.cookies.get('gh_admin');
+
+      if (!authCookie || authCookie.value !== '1') {
       // For API routes, return JSON error
       if (pathname.startsWith('/api/')) {
         return NextResponse.json(
@@ -52,12 +49,6 @@ export function middleware(request: NextRequest) {
   }
 
   return NextResponse.next();
-}
-
-function isUserAuthorized(user: User) {
-    if (user.username === 'admin') {
-        return true;
-    }
 }
 
 export const config = {
